Enhance resource handling and add Properties class
Updated `Resource.cpp` to use `float` for `CooledownTime`, enabling more precise cooldown durations. Improved clarity in `Resource.h` by explicitly referencing the `Harmony::Internals::ResourcesHandler` namespace.

Added new methods to `ResourcesHandler` in `ResourceHandler.cpp`, including `update`, `loadResource`, `unloadResource`, and `acquireResource`, to enhance resource management and thread safety.

Introduced the `Properties` class in `Properties.cpp` and `Properties.h`, inheriting from `Resource_t` and `Configuration`. Implemented `load` and `unload` methods with configuration file handling and encapsulated access to `Configuration` methods.

These changes improve precision, clarity, and extensibility in the resource management system.

+++ Harmony/Resource.cpp
@@ -6,7 +6,7 @@ namespace Harmony::Resources {
 		: configuration_(configuration), available_(false), id_(id) {
 
 		alwaysLoaded_ = configuration_.get<bool>({ "AlwaysLoaded" }).value_or(false);
-		cooledownTime_ = Time::fromSeconds(configuration_.get<int>({ "CooledownTime" }).value_or(5));
+		cooledownTime_ = Time::fromSeconds(configuration_.get<float>({ "CooledownTime" }).value_or(5));
 		required_ = configuration_.get<bool>({ "Required" }).value_or(false);
 	}
